UPSTREAM: drm/i915: Protect timeline->hwsp dereferencing
As not only is the signal->timeline volatile, so will be acquiring the
timeline's HWSP. We must first carefully acquire the timeline from the
signaling request and then lock the timeline. With the removal of the
struct_mutex serialisation of request construction, we can have multiple
timelines active at once, and so we must avoid using the nested mutex
lock as it is quite possible for both timelines to be establishing
semaphores on the other and so deadlock.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20190919111912.21631-3-chris@chris-wilson.co.uk
(cherry picked from commit 9eee0dd7d3a4e2620257f1532cba8f013aee954f)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:146438784, b:146143855, b:145022885, b:144953588
TEST=Boot to graphics on TGL board provided Mesa Iris enabled see: b:141490430
TEST=Run WebGL benchmark with 1k+ fish should not show blue fish
TEST=Font should be clear when typing on login screen

Change-Id: I700e8bf6a1c87ed25db001846f58d18d7190ba36