UPSTREAM: iio: core: Add optional symbolic label to device attributes
If a label is defined in the device tree for this device add that
to the device specific attributes. This is useful for userspace to
be able to identify an individual device when multiple identical
chips are present in the system.

Tested-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Phil Reid <preid@electromag.com.au>
Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
(cherry picked from commit 2c3d0c9ffd24d9b4c62c5dfb2104695a614be28c)

BUG=b:162279829
TEST=See label for iio devices in sysfs

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I5594572328d690351a97712fa307e492b33a1886
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2324312
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>