CHROMIUM: iwl7000: disable WRT data collection
Add a debug override option so that WRT data collection can be disabled
via iwl-dbg-cfg.ini.

This is needed for VP (a.k.a. ESL), in order to stabilize the testing
environment and to save time since the dumps are not used.

Change-Id: I87799c08eb18adf779ec71f6ca9731829721cdb6
Signed-off-by: Roee Goldfiner <roee.h.goldfiner@intel.com>
iwl7000-tree: cc707b0d9380b1815929a52efd145a0b3c15c9d8