BACKPORT: io_uring: run dependent links inline if possible
Currently any dependent link is executed from a new workqueue context,
which means that we'll be doing a context switch per link in the chain.
If we are running the completion of the current request from our async
workqueue and find that the next request is a link, then run it directly
from the workqueue context instead of forcing another switch.

This improves the performance of linked SQEs, and reduces the CPU
overhead.

Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit ba816ad61fdf31f59f423a773b00bfa2ed38243a)

Needed to apply "561fb0 io_uring: replace workqueue usage with io-wq".

Conflict:
- fs/io_uring.c (minor context diff)

BUG=chromium:901139
TEST=build

Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>
Change-Id: I1e2ce1fed8c26bfaa44b49fdd952cfbbe91bb14e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2624318
Reviewed-by: Dylan Reid <dgreid@chromium.org>