UPSTREAM: rtw88: pci: reset H2C queue indexes in a single write
If the driver doesn't reset the host's and device's indexes
in a single write, the indexes will become different in a
short period. And it will confuse the DMA engine, make it
start to process non-existed entries.

Better to Write-1-to-reset the indexes, for the DMA engine
to know that this is a reset of the H2C queue, not a kick
off.

Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
(cherry picked from commit 6f0b0d28fde849a404f4b307887405e326866e11)

BUG=b:124400293
TEST=see if there's any DMA error

Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Change-Id: Ib2f4904f0533351737b534f54375e228e2356b10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2077262
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>