UPSTREAM: rtw88: Don't set RX_FLAG_DECRYPTED if packet has no encryption
The value of GET_RX_DESC_SWDEC() indicates that if this RX
packet requires software decryption or not. And software
decryption is required when the packet was encrypted and the
hardware failed to decrypt it.

So, GET_RX_DESC_SWDEC() is negative does not mean that this
packet is decrypted, it might just have no encryption at all.
To actually see if the packet is decrypted, driver needs to
further check if the hardware has successfully decrypted it,
with a specific type of encryption algorithm.

Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
(cherry picked from commit 0649ff58a0f64df51bd24b38420da0875fbc7c00)

BUG=b:124400293
TEST=start AP with WPA2, and connect to it, should fail without this patch

Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Change-Id: I63439f1965a9d60e0daa8e3688e4c062ae9f55c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2077840
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>