UPSTREAM: drm/edid: Fix HDMI VIC handling
Extract drm_mode_hdmi_vic() to correctly calculate the final HDMI
VIC for us. Currently this is being done a bit differently between
the AVI and HDMI infoframes. Let's get both to agree on this.

We need to allow the case where a mode is both 3D and has a HDMI
VIC. Currently we'll just refuse to generate the HDMI infoframe when
we really should be setting HDMI VIC to 0 and instead enabling 3D
stereo signalling.

If the sink doesn't even support the HDMI infoframe we should
not be picking the HDMI VIC in favor of the CEA VIC, because then
we'll end up not sending either VIC in the end.

Cc: Wayne Lin <waynelin@amd.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20191004141914.20600-3-ville.syrjala@linux.intel.com
Reviewed-by: Uma Shankar <uma.shankar@intel.com>
(cherry picked from commit 949561eb85bcee10248e7da51d44a0325d5e0d1b)

BUG=b:173166465
BUG=b:172626296
TEST= Build and Boot

Signed-off-by: Ikshwaku Chauhan <ikshwaku.chauhan@amd.corp-partner.google.com>
Change-Id: I9ec4ab6744433b5412bd43e2d00191e357b3d0d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2550476
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>