UPSTREAM: drm/i915/gt: Treat a busy timeline as 'active' while waiting
If we cannot claim the timeline->mutex while preparing for a wait on it,
we have to skip the timeline. In doing so, treat it as active so that
under a intel_gt_wait_for_idle() loop, we repeat the wait after
scheduling away.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20191006165002.30312-4-chris@chris-wilson.co.uk
(cherry picked from commit a1b58ee3cb34c433167b7b7d3465ab3ca0e3daf1)

BUG=b:152719649
TEST=Test Graphics/Media/Display use cases

Signed-off-by: Ap, Kamal <kamal.ap@intel.com>
Change-Id: I34a5e56ac3b09bfc3ebede4eb9fe060a3d32ac98