UPSTREAM: drm/msm: always dump buffer base/size
Even if we are not dumping the buffer's contents, it is useful to log
their base address and size.  This makes it easier to see when different
gpu pointers point to a single buffer, for example higher mipmap levels
of a single texture.

Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit abdfd18fe07373981c5375a79a76703add0db0f0)

BUG=b:143638832
TEST=Build and boot on cheza

Change-Id: I9941f9ca9375784d73293671e4a915d27d4f7ec6
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1995838
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>