UPSTREAM: Bluetooth: hci_qca: Don't vote for specific voltage
Devices with specific voltage requirements should not request voltage
from the driver, but instead rely on the system configuration to define
appropriate voltages for each rail.

This ensures that PMIC and board variations are accounted for, something
that the 0.1V range in the hci_qca driver currently tries to address.
But on the Lenovo Yoga C630 (with wcn3990) vddch0 is 3.1V, which means
the driver will fail to set the voltage.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit f2edd66e515b9944a7e516510a54959e5004181b)

BUG=b:143638832
TEST=none

Change-Id: I31f3a4e1fa50d7a845bff7e9acaea36e7b8f9257
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1954600
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>