UPSTREAM: mac80211: Shrink the size of ack_frame_id to make room for tx_time_est
To implement airtime queue limiting, we need to keep a running account of
the estimated airtime of all skbs queued into the device. Do to this
correctly, we need to store the airtime estimate into the skb so we can
decrease the outstanding balance when the skb is freed. This means that the
time estimate must be stored somewhere that will survive for the lifetime
of the skb.

To get this, decrease the size of the ack_frame_id field to 6 bits, and
lower the size of the ID space accordingly. This leaves 10 bits for use for
tx_time_est, which is enough to store a maximum of 4096 us, if we shift the
values so they become units of 4us.

BUG=b:178754244
TEST=build and boot on asurada

(cherry picked from commit 6912daed05e1370af5253aea6f2116805c0e57f8)

Signed-off-by: Sean Wang <objelf@gmail.com>
Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>
Link: https://lore.kernel.org/r/157182474063.150713.16132669599100802716.stgit@toke.dk
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: I6a675335e9a8aac089df9bf90be57d9a654ed707
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2784517
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Claire Chang <tientzu@chromium.org>
Tested-by: Claire Chang <tientzu@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@chromium.org>