UPSTREAM: drm/i915/display/psr: Print in debugfs if PSR is not enabled because of sink
Right now if sink reported any PSR error or if it fails to
acknowledge the PSR wakeup it sets a flag and do not attempt to
enable PSR anymore. That is the safest approach to avoid repetitive
glitches and allowed us to have PSR enabled by default.

But from time to time even good PSR panels have a PSR error, causing
tests to fail. And for now we are not yet to the point were we could
try to recover from PSR errors, so lets add this information to the
debugfs so IGT can check if PSR is disabled because of sink errors or
not and eliminate this noise from CI runs.

Cc: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Cc: Matt Roper <matthew.d.roper@intel.com>
Cc: Ap Kamal <kamal.ap@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>
Reviewed-by: Ramalingam C <ramalingam.c@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20191023214932.94679-1-jose.souza@intel.com
(cherry picked from commit b7412c6b2203da6187569de600a3d2e8630f9caf)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:146438784, b:146143855, b:145022885, b:144953588
TEST=Boot to graphics on TGL board provided Mesa Iris enabled see: b:141490430
TEST=Run WebGL benchmark with 1k+ fish should not show blue fish
TEST=Font should be clear when typing on login screen

Change-Id: Ice102624b6adae879d7532a54590d5f9b3830e0f