UPSTREAM: iommu/io-pgtable-arm: Rationalise size check
It makes little sense to only validate the requested size after we think
we've found a matching block size - making the check up-front is simple,
and far more logical than waiting to walk off the bottom of the table to
infer that we must have been passed a bogus size to start with.

We're missing an equivalent check on the unmap path, so add that as well
for consistency.

Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit f7b90d2c7422a815c094961751582347935045cd)

BUG=b:150320512
TEST=Boot trogdor with wifi/modem

Signed-off-by: Evan Green <evgreen@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Icc1711c3c7af1e384573d4517cec9742e345d866
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2169160
Reviewed-by: Sean Paul <seanpaul@chromium.org>