UPSTREAM: iommu/io-pgtable-arm: Simplify PGD size handling
We use data->pgd_size directly for the one-off allocation and freeing of
the top-level table, but otherwise it serves for ARM_LPAE_PGD_IDX() to
repeatedly re-calculate the effective number of top-level address bits
it represents. Flip this around so we store the form we most commonly
need, and derive the lesser-used one instead. This cuts a whole bunch of
code out of the map/unmap/iova_to_phys fast-paths.

Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit c79278c185c8848fefc25cf304eecec9c4623a40)

BUG=b:150320512
TEST=Boot trogdor with wifi/modem

Signed-off-by: Evan Green <evgreen@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I045fe64eaef31ab110ad95c31aeb774ffd614fbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2169163
Reviewed-by: Sean Paul <seanpaul@chromium.org>