UPSTREAM: mac80211: don't re-parse elems in ieee80211_assoc_success()
We've already parsed the same data in the caller, so we can
pass it. The only thing is that we might fill in more details
in ieee80211_assoc_success(), but that doesn't bother the
caller, so it's fine to do even when we share the parsed data.

This reduces the stack space usage of the call stack here,
Arnd reported it had grown above the 1024 byte warning limit.

Reported-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Link: https://lore.kernel.org/r/20191028125240.cb7661671bd2.I757c8752bf4f2f35e54f5e0a2c0a9cd9216c3d8b@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit f61d7884cef8f1a46ed676adac313b7b53211a8a)

BUG=b:178542556
TEST=none

Signed-off-by: Zong-Zhe Yang <kevin_yang@realtek.com>
Change-Id: I757c8752bf4f2f35e54f5e0a2c0a9cd9216c3d8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2875082
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>