UPSTREAM: dma-direct: provide mmap and get_sgtable method overrides
For dma-direct we know that the DMA address is an encoding of the
physical address that we can trivially decode.  Use that fact to
provide implementations that do not need the arch_dma_coherent_to_pfn
architecture hook.  Note that we still can only support mmap of
non-coherent memory only if the architecture provides a way to set an
uncached bit in the page tables.  This must be true for architectures
that use the generic remap helpers, but other architectures can also
manually select it.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Max Filippov <jcmvbkbc@gmail.com>
(cherry picked from commit 34dc0ea6bc960f1f57b2148f01a3f4da23f87013)

BUG=b:149527406
TEST=network_WiFi_Perf.vht80

Signed-off-by: Claire Chang <tientzu@chromium.org>
Change-Id: I083830dfa728fd464b9441f9a012721e1d8af4b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2717669
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>