UPSTREAM: PM / EM: Declare EM data types unconditionally
The structs representing capacity states and performance domains of an
Energy Model are currently only defined for CONFIG_ENERGY_MODEL=y. That
makes it hard for code outside PM_EM to manipulate those structures
without a lot of ifdefery or stubbed accessors.

So, move the declaration of the two structs outside of the
CONFIG_ENERGY_MODEL ifdef. The client code (e.g. EAS or thermal) always
checks the return of em_cpu_get() before using it, so the exising code
is still safe to use as-is.

Reported-by: kbuild test robot <lkp@intel.com>
Signed-off-by: Quentin Perret <qperret@google.com>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Link: https://lore.kernel.org/r/20191030151451.7961-3-qperret@google.com

(cherry picked from commit 27a47e422ef3cb09f6a428e2b05eb79079506875)

BUG=b:170692160
TEST=none

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I339033703166b175a0758adabb70e6342000d8b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2573186
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>